requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
packaging==24.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
